use Thread::id() instead of libc::getpid()
libc::getpid() was a really bad choice - besides requiring a separate crate and
an unsafe block, it doesn't return a distinct tid. On Linux, gettid()
does this. On macOS, there's thread_selfid or pthread_threadid_np. Avoid
all that and stick to Rust's standard library. The cost is that the
output will be super ugly. Motivation to name threads, I guess.